% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{print.boot_pred}
\alias{print.boot_pred}
\title{Print a boot_pred object}
\usage{
\method{print}{boot_pred}(x, digits = 4, max_rows = 10, ...)
}
\arguments{
\item{x}{An object of class \code{"boot_pred"}.}

\item{digits}{Number of digits to display (default: 4).}

\item{max_rows}{Maximum number of statistic elements to display
(default: 10).}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a boot_pred object
}
