% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{pred_base}
\alias{pred_base}
\title{Calculate predictions (marginal expectations)}
\usage{
pred_base(eta, sigma = NA, link_name = NA, num_nodes = 15, inv_link = NA)
}
\arguments{
\item{eta}{An n-vector of linear predictor values (eta = X \%*\% beta)}

\item{sigma}{An n-vector of standard deviations for the random effects.
If NA (default), assumes no random effects (sigma = 0).}

\item{link_name}{Character string specifying the link function. Currently supports
"identity", "log", or "sqrt". Use NA if providing a custom inv_link function.}

\item{num_nodes}{Number of nodes for Gaussian quadrature used to calculate predictions
when using a custom inv_link function (default: 15)}

\item{inv_link}{Custom inverse link function. If provided, numerical integration
via Gaussian quadrature is used. Use NA if providing link_name instead.}
}
\value{
A vector of predicted (fitted) values on the natural scale
}
\description{
Calculate predictions (marginal expectations)
}
