% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{pred_base}
\alias{pred_base}
\title{Calculate predictions (marginal expectations)}
\usage{
pred_base(X, Beta, sigma, link = "log", num_nodes = 15)
}
\arguments{
\item{X}{An n x p matrix of predictors}

\item{Beta}{A p-vector of regression coefficients}

\item{sigma}{An n-vector of standard deviations for the latent variables,}

\item{link}{The inverse link function, currently one of
"Linear", "Exponential", or "Logistic"}

\item{num_nodes}{Number of nodes for Gaussian quadrature used to calculate predictions}
}
\value{
A vector of predicted (fitted) values
}
\description{
Calculate predictions (marginal expectations)
}
