% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{print.boot_pred}
\alias{print.boot_pred}
\title{Print method for bootstrap confidence intervals}
\usage{
\method{print}{boot_pred}(x, digits = 4, max_rows = 10, ...)
}
\arguments{
\item{x}{A \code{boot_pred} object from \code{\link{boot_pred}}.}

\item{digits}{Number of digits displayed. Default 4.}

\item{max_rows}{Maximum number of rows displayed. Default 10.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for bootstrap confidence intervals
}
